Add game loop that updates the game state based on events collected in between renders and renders the new game state. The loop is scheduled with `requestAnimationFrame`.
The inputEvents function converts domEvents to events related to the game. Those events are
handled in the game loop.

The update function takes in an event to be handled and a game state and outputs a new game state that was modified based on the event.

Add "ramda" library for general functional programming tools.

Add "immutable.js" library for efficient immutable data structures.

Add "rx.js" library for the observable data structure.